move message list to it's own store
that is thrown away and reacreated on selecting a chat,
so selecting a chat auto fixes possible problems again,
also makes our code simpler
